receipt sizing and rev positioning
